refactor(database) - clean up database files/structure
